pyproject.toml
[tool.poetry]
name = "immo-tools-cli"
version = "0.1.0"
description = "compute loan home cost in cli"
license = "MIT"
authors = ["thomas.marquis.dev <thomas.marquis.dev@gmail.com>"]
repository = "https://github.com/thomas-marquis/immo-tools-cli"
homepage = "https://github.com/thomas-marquis/immo-tools-cli"
packages = [
    {include = "immo_cli"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
immo-tools-lib = "0.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
